made Framerate::millisecondsPerSecond into a float since maths involving it mostly uses floats anyway.

+++ Framerate.h
@@ -1,7 +1,7 @@
 #pragma once
 namespace Framerate
 {
-	constexpr unsigned long millisecondsPerSecond = 1000;
+	constexpr float millisecondsPerSecond = 1000.0;
 	constexpr unsigned long maxFPS = 30;
 	constexpr unsigned long minimumFrameTime = millisecondsPerSecond / maxFPS;
 }
